requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
crewai==0.159.0
crewai-tools==0.62.0
python-dotenv
pypdf==5.9.0
